Add files via upload
A natural number is given. It is required to determine whether the year with the given number is a leap year. Print the text "YES", otherwise you can "NO". Recall that according to the Gregorian calendar, a year is a leap year if its number is a multiple of 4, but not a multiple of 100, or if it is a multiple of 400.