Add files via upload
The chess king moves horizontally, vertically and diagonally, but only 1 square. Given two different cellscheckerboard, determine if the king can get from the first cell to the second in one move.The program receives as input four numbers from 1 to 8 each, specifying the column number and row number first for the firstcells, then for the second cell.